+++ packages/tools/src/utilities/rtstruct/index.ts
@@ -5,23 +5,10 @@ import { generateContourSetsFromLabelmap } from './generateContourSetsFromLabelm
 
 import AnnotationToPointData from './AnnotationToPointData';
 
-import getPatientModule from './getPatientModule';
-import getReferencedFrameOfReferenceSequence from './getReferencedFrameOfReferenceSequence';
-import getReferencedSeriesSequence from './getReferencedSeriesSequence';
-import getRTROIObservationsSequence from './getRTROIObservationsSequence';
-import getRTSeriesModule from './getRTSeriesModule';
-import getStructureSetModule from './getStructureSetModule';
-
 export {
   contourFinder,
   mergePoints,
   detectContourHoles,
   generateContourSetsFromLabelmap,
   AnnotationToPointData,
-  getPatientModule,
-  getReferencedFrameOfReferenceSequence,
-  getReferencedSeriesSequence,
-  getRTROIObservationsSequence,
-  getRTSeriesModule,
-  getStructureSetModule,
 };
